Move attached queue tracking to Message object
